teams/cloud/container: Use 'toolbx' as VARIANT_ID in os-release(5)
The same change was made to the Kickstart equivalent of the
fedora-toolbox:41 OCI image recently [1].

This is meant to distinguish OCI containers and images that are designed
specifically for Toolbx from others.  Toolbx containers are long-lasting
pet containers for interactive command line use, which makes them
substantially different from short-lived containers running services.

Therefore, it can be useful to be able to identify Toolbx containers and
images when generating statistics about Fedora usage.

[1] fedora-kickstarts commit 0d99c64eb2721c5b
    https://pagure.io/fedora-kickstarts/c/0d99c64eb2721c5b
    https://pagure.io/fedora-kickstarts/pull-request/1015

https://pagure.io/Fedora-Council/tickets/issue/449