Keep kbd-misc for now
On ppc64le, power-utils is pulled in by being default in Core group.
This in turn pulls in power-utils-core, which pulls in systemd-udev.
When kiwi goes to remove kbd-misc on ppc64le only, it fails because
systemd-udev is a protected package. On other arches since it's not
installed, it works.

So, we are going to just drop this for now and revisit solutions after
Beta is out the door.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>